More import cleanup for ServerModels refactor
