+++ app/src/main/java/com/bookyrself/bookyrself/views/EventInvitesFragment.java
@@ -17,7 +17,7 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 
 import com.bookyrself.bookyrself.R;
-import com.bookyrself.bookyrself.data.ResponseModels.EventDetail.EventDetail;
+import com.bookyrself.bookyrself.data.ServerModels.EventDetail.EventDetail;
 import com.bookyrself.bookyrself.presenters.EventInvitesFragmentPresenter;
 import com.firebase.ui.auth.AuthUI;
 import com.google.firebase.auth.FirebaseAuth;
